+++ projectionizer/tests/test_sscx.py
@@ -33,41 +33,11 @@ def test_assign_synapse_fiber():
     assert ret.equals(pd.DataFrame({'sgid': [0, 2, 1]}))
 
 
-def test_column_layers():
-    layer_starts, layer_thickness = sscx.column_layers(RAT_LAYERS)
-    assert layer_starts == {'L6': 0, 'L4': 1225, 'L5': 700, 'L2': 1766, 'L3': 1414, 'L1': 1914}
-    assert layer_thickness == {'L6': 700, 'L4': 189, 'L5': 525, 'L2': 148, 'L3': 352, 'L1': 164}
-
-
-def test_recipe_to_height_and_density():
-    profile = [[0.05, 0.01], [0.15, 0.02], [0.25, 0.03], [0.35, 0.04], [0.45, 0.04],
-               [0.55, 0.04], [0.65, 0.03], [0.75, 0.02], [0.85, 0.01], [0.95, 0.01]]
-
-    res = sscx.recipe_to_height_and_density(RAT_LAYERS, 'L4', 0, 'L3', 0.5, profile)
-    assert_allclose(res,
-                    [(1225.0, 0.01),
-                     (1261.5, 0.02),
-                     (1298.0, 0.03),
-                     (1334.5, 0.04),
-                     (1371.0, 0.04),
-                     (1407.5, 0.04),
-                     (1444.0, 0.03),
-                     (1480.5, 0.02),
-                     (1517.0, 0.01),
-                     (1553.5, 0.01),
-                     (1590.0, 0.01)])
-
-
 def test_relative_recipe_to_height_and_density():
     profile = [[0.05, 0.01], [0.15, 0.02], [0.25, 0.03], [0.35, 0.04], [0.45, 0.04],
                [0.55, 0.04], [0.65, 0.03], [0.75, 0.02], [0.85, 0.01], [0.95, 0.01]]
 
-    layers = [[6, 0],
-              [5, 0],
-              [4, 0],
-              [3, 0],
-              [2, 0],
-              [1, 0]]
+    layers = [*np.arange(6, 0, -1)]
 
     layer4 = 4
     layer3 = 3
@@ -103,7 +73,7 @@ def test_recipe_to_relative_heights_per_layer():
     # min limit is 0, max limit is the highest point of L6 in [PH]y.nrrd
     atlas = Atlas.open(TEST_DATA_DIR)
     distance = atlas.load_data('[PH]y')
-    layers = [(6, None)]
+    layers = ['L6']
 
     ret = sscx.recipe_to_relative_heights_per_layer(distance, atlas, layers)